# Improve PIL and COLOR_BGR context image decode performance