@vinnamkim 2024-05-17
Rebased and targeted to `releases/1.7.0`